implemented adding trips and trips details

+++ app/src/main/java/com/mrdor1stan/buonjourney/data/DatabaseRepository.kt
@@ -1,8 +1,5 @@
 package com.mrdor1stan.buonjourney.data
 
-import androidx.room.Delete
-import androidx.room.Query
-import androidx.room.Transaction
 import com.mrdor1stan.buonjourney.data.db.BuonjourneyDao
 import com.mrdor1stan.buonjourney.data.db.PlaceDto
 import com.mrdor1stan.buonjourney.data.db.TripDto
@@ -12,7 +9,7 @@ import kotlinx.coroutines.flow.Flow
 class DatabaseRepository(private val dao: BuonjourneyDao) {
     suspend fun addTrip(trip: TripDto) = dao.addTrip(trip)
     fun getTrips(): Flow<List<TripsDetailsDto>> = dao.getTrips()
-    fun getTrip(id: Int): Flow<TripsDetailsDto> = dao.getTrip(id)
+    fun getTrip(id: Long): Flow<TripsDetailsDto> = dao.getTrip(id)
     suspend fun deleteTrip(trip: TripDto) = dao.deleteTrip(trip)
     suspend fun deleteTrip(id: Int) = dao.deleteTrip(id)
     suspend fun addPlace(place: PlaceDto) = dao.addPlace(place)
